+++ Code/projects/observer/src/Monitor.hpp
@@ -8,7 +8,7 @@ public:
     Monitor() {};
     virtual ~Monitor() {};
     virtual void start(int pid) = 0;
-    virtual std::pair<long, long> stop() = 0;
+    virtual std::pair<double, double> stop() = 0;
 };
 
 #endif // MONITOR_H
